\name{hygiene}
\alias{hygiene}
\docType{data}
\title{
Gingival epithelial cell line response to oral pathogen infections - ExpressionSet Object
}
\description{
This is the ExpressionSet object for the Gingival epithelial cell line response to oral pathogen infections.
}
\usage{data(hygiene)}
\format{
  A data frame with 22286 observations on the following 12 variables.
  \describe{
    \item{\code{GSM245725}}{a numeric vector}
    \item{\code{GSM245726}}{a numeric vector}
    \item{\code{GSM245727}}{a numeric vector}
    \item{\code{GSM245728}}{a numeric vector}
    \item{\code{GSM245729}}{a numeric vector}
    \item{\code{GSM245730}}{a numeric vector}
    \item{\code{GSM245731}}{a numeric vector}
    \item{\code{GSM245732}}{a numeric vector}
    \item{\code{GSM245733}}{a numeric vector}
    \item{\code{GSM245734}}{a numeric vector}
    \item{\code{GSM245735}}{a numeric vector}
    \item{\code{GSM245736}}{a numeric vector}
  }
}
\examples{
data(hygiene)
}
\keyword{datasets}
