\name{OralHygiene-package}
\alias{OralHygiene-package}
\alias{OralHygiene}
\docType{package}
\title{
Gingival epithelial cell line response to oral pathogen infections
}
\description{
This dataset was obtained from the National Center for Biotechnology Information.  12 samples were taken:  4 HIGK cells infected with porphyromonas gingivalis, 4 HIGK cells infected with aggregatibacter actinomycetemcomitans, and 4 HIGK cells infected with sham controls.}
\details{
\tabular{ll}{
Package: \tab OralHygiene\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-02-22\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}
The object 'hygiene' in this library contains gene expression data.  See help(hygiene) for details.
}
\author{
Jarrod E. Dalton

Maintainer: Jarrod E. Dalton <jarrod.dalton@gmail.com>
}
\keyword{ package }
